requirements.txt
# ----------------------------------------------------------------------
# NOC Python requirements
# Following extras can be set:
# * activator
# * cache-redis
# * cache-memcached
# * classifier
# * cython
# * dev
# * login-ldap
# * login-pam
# * login-radius
# * prod-tools
# * supervisor
# * testing
# ----------------------------------------------------------------------
# Copyright (C) 2007-2020 The NOC Project
# See LICENSE for details
# ----------------------------------------------------------------------

# Basic node requirements
Django==3.0.6
FastAPI==0.61.1
Jinja2==2.8
PyYAML==5.3.1
Pygments==2.2.0
XlsxWriter==0.9.3
atomicl==0.1.1
blinker==1.3
bsdiff4==1.1.9
cachetools==4.1.0
crontab==0.22.6
csiphash==0.0.5
demjson==2.2.4
fs==2.4.11
future==0.17.1
geojson==2.5.0
geopy==0.97
grpcio==1.32.0
http-parser==0.9.0
jsonschema==2.4.0
manhole==1.6.0
markerlib==0.6.0
mongoengine==0.20.0
networkx==2.4
numpy==1.18.3
orjson==3.3.1
progressbar2==3.51.0
protobuf==3.13.0
psycopg2==2.8.5
pydantic==1.6.1
pymongo==3.11.0
pynsq==0.9.0
pyproj==2.6.0
python-consul==1.1.0
python-creole==1.3.2
python-dateutil==2.4.0
python-jose[cryptography]==3.2.0
python-multipart==0.0.5
pytz==2020.1
requests==2.20.0
setproctitle==1.1.10
starlette==0.13.8
tornado==6.0.4
typing-extensions==3.7.4.3; python_version < '2.8'
uvicorn==0.11.8
yappi==1.2.5
# Activator requirements
ssh2-python==0.18.0.post1; extra == "activator"
# Classifier
bitarray==0.8.1; extra == "classifier"
# Cache-Redis
hiredis==1.0.1; extra == "cache-redis"
redis==3.5.2; extra == "cache-redis"
# Cache-Memcached
pylibmc==1.5.1; extra == "cache-memcached"
# Dev requirements
Babel==2.3.4; extra == "dev"
ipython==7.14.0; extra == "dev"
pojson==0.6; extra == "dev"
# Login LDAP
ldap3==2.6.1; extra == "login-ldap"
# Login PAM
python-pam==1.8.4; extra == "login-pam"
# Login RADIUS
pyrad==2.0; extra == "login-radius"
# Standalone node with supervisor
supervisor==4.1.0; extra == "supervisor"
supervisor-serialrestart==0.1.1; extra == "supervisor"
# Additional prod tools
alerta==4.7.16; extra == "prod-tools"
sentry-sdk==0.16.5; extra == "prod-tools"
supervisor-wildcards==0.1.3; extra == "prod-tools"
# Test
Coverage[toml]==5.0.3; extra == "testing"
pytest==5.4.1; extra == "testing"
pytest-assume==2.2.1; extra == "testing"
pytest-dependency==0.5.1; extra == "testing"
# Cython
Cython==0.29.20; extra == "cython"
